pyproject.toml
[project]
name = "multioutreg"
version = "0.0.1"
description = "Multi-Output Regressors"
readme = "README.md"
requires-python = ">=3.8"
# authors = [
#     {name = "takotime808", email = "takotime808"},
# ]
license = {file = "LICENSE"}

dependencies = [
  "typer-cli", # 0.12.0
  "wasabi",
  "termcolor",
  "pyfiglet",
  "terminology",
  "numpy>=1.22,<2.0",
  "pandas>=1.5,<2.2",
  "matplotlib>=3.5,<4.0",
  "scikit-learn>=1.2,<1.5",
  "uncertainty-toolbox",#>=0.1.4,<0.1.7",
]

[tool.setuptools.packages.find]
where = [".", "mor_cli", "multioutreg"]
namespaces = false

[project.scripts]
multioutreg = "mor_cli:main"

[build-system]
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"

# This field adds keywords for your project which will appear on the
# project page. What does your project relate to?
#
# Note that this is a list of additional keywords, separated
# by commas, to be used to assist searching for the distribution in a
# larger catalog.
# keywords = ["sample", "setuptools", "development"]

# [project.urls]
# "Homepage" = "https://github.com/pypa/sampleproject"
# "Bug Reports" = "https://github.com/pypa/sampleproject/issues"
# "Funding" = "https://donate.pypi.org"
# "Say Thanks!" = "http://saythanks.io/to/example"
# "Source" = "https://github.com/pypa/sampleproject/"
